requirements.txt
streamlit
python-dotenv
langchain-google-genai
langchain-core
Pillow
fastapi
uvicorn
python-dotenv
langchain-google-genai
pydantic
python-multipart
streamlit
requests
Pillow